chore: renaming and minor improvements
