+++ app/modules/intelligence/agents_copy/chat_agents/crewai_agent.py
@@ -32,7 +32,7 @@ class AgentConfig(BaseModel):
     max_iter: int = 15
 
 
-class CrewAIRagAgent(ChatAgent):
+class CrewAIAgent(ChatAgent):
     def __init__(
         self,
         llm_provider: ProviderService,
